Updated typedefs

+++ src/raspberrypi/os.h
@@ -103,9 +103,9 @@
 //
 //---------------------------------------------------------------------------
 typedef unsigned char BYTE;
-typedef unsigned short WORD;
-typedef unsigned long DWORD;
-typedef unsigned long long QWORD;
+typedef uint16_t WORD;
+typedef uint32_t DWORD;
+typedef uint64_t QWORD;
 typedef int BOOL;
 typedef char TCHAR;
 typedef char *LPTSTR;
